feat: use bindings
