+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cfd-js",
-  "version": "0.1.10",
+  "version": "0.1.11",
   "domain": "cryptogarageinc",
   "repository": {
     "type": "git",
@@ -157,7 +157,7 @@
   "dependencies": {
     "@types/node": "^13.11.0",
     "big-integer": "^1.6.48",
-    "bindings": "*",
+    "bindings": "^1.5.0",
     "cmake-js": "^5.3.2",
     "del-cli": "^3.0.0",
     "got": "^10.7.0",
